+++ NEWS.md
@@ -1,3 +1,5 @@
+# riem (development version)
+
 # riem 0.3.2
 
 * Remove last usage of vcr as the choice was made to use httptest2 instead.