+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * New arguments `data`, `latlon`, `report_type`, `elev` (#48, @JElchison).
 
-* Breaking change: `laton` default to `FALSE`: you need to explicitly set it to
+* Breaking change: `latlon` default to `FALSE`: you need to explicitly set it to
   `TRUE` for the latitude and longitude of the station 
   to be included in the output.