+++ NEWS.md
@@ -1,4 +1,4 @@
-# riem (development version)
+# riem 1.0.0
 
 * Breaking change: `date_start` and `station` no longer have default values.